contributors/adobe/submitted/regions/webkit-flow-into-parsingTest.html

<!DOCTYPE html>
<html>
    <head>
        <title>CSS Regions Parsing Test: Parse webkit-flow-into property</title>
        <link rel="author" title="Jacob Goldstein" href="mailto:jacobg@adobe.com"/>
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#properties-and-rules"/>
        <meta name="flags" content="" />
        <meta name="assert" content="Value specified for the webkit-flow-into property should be parsed correctly."/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>        
    </head>
    
    <body>
        <div id="log"></div>
        
        <script type="text/javascript">
            function testParse(declaration) {
                var div = document.createElement("div");
                div.setAttribute("style", declaration);
                return div.style.webkitFlowInto;
            }
            
            function testComputedStyle(value) {
                var div = document.createElement("div");
                document.body.appendChild(div);
                div.style.setProperty("-webkit-flow-into", value);
                var webkitFlowComputedValue = getComputedStyle(div).getPropertyValue("-webkit-flow-into");
                document.body.removeChild(div);
                return webkitFlowComputedValue;
            }

            function testNotInherited(parentValue, childValue) {
                var parentDiv = document.createElement("div");
                document.body.appendChild(parentDiv);
                parentDiv.style.setProperty("-webkit-flow-into", parentValue);
                var childDiv = document.createElement("div");
                parentDiv.appendChild(childDiv);
                childDiv.style.setProperty("-webkit-flow-into", childValue);
                var childWebKitFlowComputedValue = getComputedStyle(childDiv).getPropertyValue("-webkit-flow-into");
                parentDiv.removeChild(childDiv);
                document.body.removeChild(parentDiv);
                return childWebKitFlowComputedValue;
            }

            test(function() {assert_equals(testParse("-webkit-flow-into: none"), "none")}, "testParse: Assigned none to property, expected return = none"); 
            test(function() {assert_equals(testParse("-webkit-flow-into: first-flow"), "first-flow")}, "testParse: Assigned first-flow to property, expected return = first-flow");
            test(function() {assert_equals(testParse("-webkit-flow-into: \'first flow\'"), "")}, "testParse: Assigned 'first-flow' to property, expected return = ''");
            test(function() {assert_equals(testParse("-webkit-flow-into: ;"), "")}, "testParse: Assigned ; to property, expected return = ''");
            test(function() {assert_equals(testParse("-webkit-flow-into: 1"), "")}, "testParse: Assigned 1 to property, expected return = ''");
            test(function() {assert_equals(testParse("-webkit-flow-into: 1.2"), "")}, "testParse: Assigned 1.2 to property, expected return = ''");
            test(function() {assert_equals(testParse("-webkit-flow-into: -1"), "")}, "testParse: Assigned -1 to property, expected return = ''");
            test(function() {assert_equals(testParse("-webkit-flow-into: 12px"), "")}, "testParse: Assigned 12px to property, expected return = ''");

            test(function() {assert_equals(testComputedStyle("none"), "none")}, "testComputedStyle: Assigned none to property, expected return = none");
            test(function() {assert_equals(testComputedStyle(""), "none")}, "testComputedStyle: Assigned \"\" to property, expected return = none");
            test(function() {assert_equals(testComputedStyle("\'first-flow\'"), "none")}, "testComputedStyle: Assigned 'first-flow' to property, expected return = none");
            test(function() {assert_equals(testComputedStyle("first-flow"), "first-flow")}, "testComputedStyle: Assigned first-flow to property, expected return = first-flow");
            test(function() {assert_equals(testComputedStyle("12px"), "none")}, "testComputedStyle: Assigned 12px to property, expected return = none");

            test(function() {assert_equals(testNotInherited("none", "none"), "none")}, "testNotInherited: Assigned none, none to property, expected return = none");
            test(function() {assert_equals(testNotInherited("none", "child-flow"), "child-flow")}, "testNotInherited: Assigned none, child-flow to property, expected return = child-flow");
            test(function() {assert_equals(testNotInherited("parent-flow", "none"), "none")}, "testNotInherited: Assigned parent-flow, none to property, expected return = none");
            test(function() {assert_equals(testNotInherited("parent-flow", "child-flow"), "child-flow")}, "testNotInherited: Assigned parent-flow, child-flow to property, expected return = child-flow");
        </script>
    </body>
</html>
